Started Folders Support
started the support for folders, the setting for googledrivedir should now
properly import and when an upload is initiated, should put all of the files
inside of that master directory.

Now creates it's own subfolder with the title of the torrent as the folder
name in order to have some organization. (It was just dumping everything into
the root folder previously)

Created get_FilesDict function to return a dictionary of the torrents contents
in order to facilicate preservation of file structures. (Not finished yet...)

Added uploadToGoogleDrive function to directly handle the uploading of the
files to google drive without having to use PyDrive, PyDrive package
should be able to be removed from the package dependencies as soon as I
clean it up a bit more.