Email overhaul + Commandline implimentation
Big overhaul here, and the code needs a bit of cleaning before I
merge with existing code.

Email settings have been added to the settings.py file in order to
allow the user to select what things they want in the table to be
sent with.  W3C Validation is intact. Not like it matters, but it's
there.

The email parameters list includes what I currently run in my table.
There are many *many* others that can be added to the list as well
if it's something anyone wants to change.

The actual uploading of files can be done inside the script as intended
and, moreso, the script recreates the file structure for the torrent's
files on Google Drive thereby adding a bit of organization.

For the moment, I still haven't yet figured out how to get around the
single file problem, since as it stands, it chooses what files to upload
based on what it receieves from Transmission-Daemon when it's run in the
form of it's env vars passed. It generates a directory that the torrent
files themselves are inside by joining bt_dir and bt_name, since most
torrents follow this proceedure of creating a directory with it's name,
so a single os.path.join later and off we goto the races.

What I suspect will happen in the case of a single file (where it
doesn't create a directory) is it'll either crash out entirely, or,
end up uploading your entire downloads folder to Google Drive.
Working on that is one of my next steps.

The script also keeps a hold of all of the responses from the Google
API as JSON entries, compiles those entries into a list, and then
parses that list as per the settings. It creates a table with headers
for each entry it finds, then adds a line for each file it uploads,
skipping folders.

PyDrive is no longer used whatsoever anymore so other than the Google
Drive API Python Package itself, there shouldn't be any need for any
3rd party packages after this point. I still have to clean up the code
to reflect that, but there aren't anymore calls to PyDrive anymore.

oauth2client support still remains largely unchanged, use of this script
overall still requires a Service Account though that will change in the
very near future. In any case, Merry Christmas.