Added many new settings.
Again, lacking on the documentation, apologies.

Most of the changes were to make it easier for the user to change various settings
in the script in order to facilitate easy of use, the CategoriesDictSetting being
the exception to that.

I'll be adding a bit of info about that and the Rules.py in a wiki coming up soon.