Moved over to settings.json
This current branch is extremely unstable. Be warned.

All settings have moved from a settings.py set of variables over to a JSON
file for ease of portability as I expand this to other languages.

Currently in test is the SetupUploader.py script that will enable the user
to setup the script using easily read instructructions and test various
aspects like authentication as it moves along. Currently, it's only set up
to make the first authorized called to the API that should list the
last 10 files that the user touched on their own google drive.

Documentation still needs to be updated for this as well.